Migrate all atomic directory tests.
These were left in test_common from a previous refactoring.
